HW02 optional

+++ src/main/java/ru/javawebinar/topjava/repository/mock/InMemoryMealRepositoryImpl.java
@@ -1,16 +1,24 @@
 package ru.javawebinar.topjava.repository.mock;
 
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
+import org.springframework.stereotype.Repository;
 import ru.javawebinar.topjava.model.Meal;
 import ru.javawebinar.topjava.repository.MealRepository;
+import ru.javawebinar.topjava.util.DateTimeUtil;
 import ru.javawebinar.topjava.util.MealsUtil;
-import ru.javawebinar.topjava.util.exception.NotFoundException;
 
+
+import java.time.LocalDate;
+import java.time.LocalDateTime;
 import java.util.*;
 import java.util.concurrent.ConcurrentHashMap;
 import java.util.concurrent.atomic.AtomicInteger;
 import java.util.stream.Collectors;
 
+@Repository
 public class InMemoryMealRepositoryImpl implements MealRepository {
+    private static final Logger log= LoggerFactory.getLogger(InMemoryMealRepositoryImpl.class);
     private Comparator<Meal> mealComparator = Comparator.comparing(Meal::getDateTime).reversed();
     private Map<Integer, Meal> repository = new ConcurrentHashMap<>();
     private AtomicInteger counter = new AtomicInteger(0);
@@ -21,14 +29,12 @@ public class InMemoryMealRepositoryImpl implements MealRepository {
 
     @Override
     public Meal save(Meal meal, int userId) {
-
-        if (meal.getUserId() != userId) {
-            return null;
-        }
+        log.info("save meal {} and userId{}", meal, userId);
 
         if (meal.isNew()) {
             meal.setId(counter.incrementAndGet());
         }
+
         repository.put(meal.getId(), meal);
         return meal;
     }
@@ -36,20 +42,43 @@ public class InMemoryMealRepositoryImpl implements MealRepository {
     @Override
     public boolean delete(int id, int userId) {
         Meal meal = repository.get(id);
+        log.info("delete by id {} and userId{}", id, userId);
+
+        if (meal == null) {
+            return false;
+        }
+
         return userId == meal.getUserId() && repository.remove(id) != null;
     }
 
     @Override
     public Meal get(int id, int userId) {
         Meal meal = repository.get(id);
-        return userId == meal.getUserId() ? repository.get(id) : null;
+        log.info("get by id {} and userId{}", id, userId);
+
+        if (meal == null || userId != meal.getUserId()) {
+            return null;
+        }
+
+        return meal;
     }
 
     @Override
     public Collection<Meal> getAll(int userId) {
-        List<Meal> meals = new ArrayList<>(repository.values());
-        meals = meals.stream().filter(meal -> meal.getUserId() == userId).sorted(mealComparator).collect(Collectors.toList());
-        return meals;
+        log.info("getAll");
+        return repository.values().stream()
+                .filter(meal -> meal.getUserId() == userId)
+                .sorted(mealComparator)
+                .collect(Collectors.toList());
+
     }
+
+    public Collection<Meal> getAllFiltered(int userId, LocalDate start, LocalDate end) {
+        log.info("getAll filtered by startDate {} and endDate{}", start, end);
+        return getAll(userId).stream()
+                             .filter(meal -> DateTimeUtil.isBetween((meal.getDate()), start, end))
+                             .collect(Collectors.toCollection(ArrayList::new));
+    }
+
 }
